finish

+++ init.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS currency (
     Id SERIAL PRIMARY KEY,
     name VARCHAR(100),
-    value INTEGER,
+    value FLOAT,
     created_at TIMESTAMP
 );
